requirements.txt
# Core Dependencies
playwright>=1.40.0
httpx>=0.25.0
aiohttp>=3.9.0

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
click>=8.1.0
loguru>=0.7.0
rich>=13.0.0
aiofiles>=23.0.0
inquirer>=3.1.0

# Security Testing
dnspython>=2.4.0
cryptography>=41.0.0
pyjwt>=2.8.0
keyring>=24.0.0

# Reporting
jinja2>=3.1.0
reportlab>=4.0.0
openpyxl>=3.1.0
matplotlib>=3.8.0

# WebSocket Testing
websockets>=12.0

# Image Processing (for visual testing)
pillow>=10.0.0
numpy>=1.24.0

# Cache
aioredis>=2.0.1
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0

# Additional
python-multipart>=0.0.6
fastapi>=0.104.0
uvicorn>=0.24.0
